cb.evaluation.log {xgboost}	R Documentation

Callback closure for logging the evaluation history

Description

Callback closure for logging the evaluation history

Usage

cb.evaluation.log()

Details

This callback function appends the current iteration evaluation results bst_evaluation available in the calling parent frame to the evaluation_log list in a calling frame.

The finalizer callback (called with finalize = TURE in the end) converts the evaluation_log list into a final data.table.

The iteration evaluation result bst_evaluation must be a named numeric vector.

Note: in the column names of the final data.table, the dash '-' character is replaced with the underscore '_' in order to make the column names more like regular R identifiers.

Callback function expects the following values to be set in its calling frame: evaluation_log, bst_evaluation, iteration.

See Also

callbacks

[Package xgboost version 1.0.0.1 Index]