cb.print.evaluation {xgboost}	R Documentation

Callback closure for printing the result of evaluation

Description

Callback closure for printing the result of evaluation

Usage

cb.print.evaluation(period = 1, showsd = TRUE)

Arguments

period
results would be printed every number of periods

showsd
whether standard deviations should be printed (when available)

Details

The callback function prints the result of evaluation at every period iterations. The initial and the last iteration's evaluations are always printed.

Callback function expects the following values to be set in its calling frame: bst_evaluation (also bst_evaluation_err when available), iteration, begin_iteration, end_iteration.

See Also

callbacks

[Package xgboost version 1.0.0.1 Index]