cb.reset.parameters {xgboost}	R Documentation

Callback closure for resetting the booster's parameters at each iteration.

Description

Callback closure for resetting the booster's parameters at each iteration.

Usage

cb.reset.parameters(new_params)

Arguments

new_params
a list where each element corresponds to a parameter that needs to be reset. Each element's value must be either a vector of values of length nrounds to be set at each iteration, or a function of two parameters learning_rates(iteration, nrounds) which returns a new parameter value by using the current iteration number and the total number of boosting rounds.

Details

This is a "pre-iteration" callback function used to reset booster's parameters at the beginning of each iteration.

Note that when training is resumed from some previous model, and a function is used to reset a parameter value, the nrounds argument in this function would be the the number of boosting rounds in the current training.

Callback function expects the following values to be set in its calling frame: bst or bst_folds, iteration, begin_iteration, end_iteration.

See Also

callbacks

[Package xgboost version 1.0.0.1 Index]