cb.save.model {xgboost}	R Documentation

Callback closure for saving a model file.

Description

Callback closure for saving a model file.

Usage

cb.save.model(save_period = 0, save_name = "xgboost.model")

Arguments

save_period
save the model to disk after every save_period iterations; 0 means save the model at the end.

save_name
the name or path for the saved model file. It can contain a sprintf formatting specifier to include the integer iteration number in the file name. E.g., with save_name = 'xgboost_ the file saved at iteration 50 would be named "xgboost_0050.model".

Details

This callback function allows to save an xgb-model file, either periodically after each save_period's or at the end.

Callback function expects the following values to be set in its calling frame: bst, iteration, begin_iteration, end_iteration.

See Also

callbacks

[Package xgboost version 1.0.0.1 Index]