dim.xgb.DMatrix {xgboost}	R Documentation

Dimensions of xgb.DMatrix

Description

Returns a vector of numbers of rows and of columns in an xgb.DMatrix.

Usage

## S3 method for class 'xgb.DMatrix'
dim(x)

Arguments

x
Object of class xgb.DMatrix

Details

Note: since nrow and ncol internally use dim, they can also be directly used with an xgb.DMatrix object.

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train
dtrain <- xgb.DMatrix(train$data, label=train$label)

stopifnot(nrow(dtrain) == nrow(train$data))
stopifnot(ncol(dtrain) == ncol(train$data))
stopifnot(all(dim(dtrain) == dim(train$data)))

[Package xgboost version 1.0.0.1 Index]