getinfo {xgboost}	R Documentation

Get information of an xgb.DMatrix object

Description

Get information of an xgb.DMatrix object

Usage

getinfo(object, ...)

## S3 method for class 'xgb.DMatrix'
getinfo(object, name, ...)

Arguments

object
Object of class xgb.DMatrix

...
other parameters

name
the name of the information field to get (see details)

Details

The name field can be one of the following:

label: label Xgboost learn from ;

weight: to do a weight rescale ;

base_margin: base margin is the base prediction Xgboost will boost from ;

nrow: number of rows of the xgb.DMatrix.

group can be setup by setinfo but can't be retrieved by getinfo.

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train
dtrain <- xgb.DMatrix(train$data, label=train$label)

labels <- getinfo(dtrain, 'label')
setinfo(dtrain, 'label', 1-labels)

labels2 <- getinfo(dtrain, 'label')
stopifnot(all(labels2 == 1-labels))

[Package xgboost version 1.0.0.1 Index]