prepare.ggplot.shap.data {xgboost}	R Documentation

Combine and melt feature values and SHAP contributions for sample observations.

Description

Conforms to data format required for ggplot functions.

Usage

prepare.ggplot.shap.data(data_list, normalize = FALSE)

Arguments

data_list
List containing 'data' and 'shap_contrib' returned by xgb.shap.data().

normalize
Whether to standardize feature values to have mean 0 and standard deviation 1 (useful for comparing multiple features on the same plot). Default FALSE.

Details

Internal utility function.

Value

A data.table containing the observation ID, the feature name, the feature value (normalized if specified), and the SHAP contribution value.

[Package xgboost version 1.7.3.1 Index]