setinfo {xgboost}	R Documentation

Set information of an xgb.DMatrix object

Description

Set information of an xgb.DMatrix object

Usage

setinfo(object, ...)

## S3 method for class 'xgb.DMatrix'
setinfo(object, name, info, ...)

Arguments

object
Object of class "xgb.DMatrix"

...
other parameters

name
the name of the field to get

info
the specific field of information to set

Details

The name field can be one of the following:

label: label Xgboost learn from ;

weight: to do a weight rescale ;

base_margin: base margin is the base prediction Xgboost will boost from ;

group: number of rows in each group (to use with rank:pairwise objective).

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train
dtrain <- xgb.DMatrix(train$data, label=train$label)

labels <- getinfo(dtrain, 'label')
setinfo(dtrain, 'label', 1-labels)
labels2 <- getinfo(dtrain, 'label')
stopifnot(all.equal(labels2, 1-labels))

[Package xgboost version 1.0.0.1 Index]