xgb.DMatrix {xgboost}	R Documentation

Construct xgb.DMatrix object

Description

Construct xgb.DMatrix object from either a dense matrix, a sparse matrix, or a local file. Supported input file formats are either a libsvm text file or a binary file that was created previously by xgb.DMatrix.save).

Usage

xgb.DMatrix(data, info = list(), missing = NA, silent = FALSE, ...)

Arguments

data
a matrix object (either numeric or integer), a dgCMatrix object, or a character string representing a filename.

info
a named list of additional information to store in the xgb.DMatrix object. See setinfo for the specific allowed kinds of

missing
a float value to represents missing values in data (used only when input is a dense matrix). It is useful when a 0 or some other extreme value represents missing values in data.

silent
whether to suppress printing an informational message after loading from a file.

...
the info data could be passed directly as parameters, without creating an info list.

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train
dtrain <- xgb.DMatrix(train$data, label=train$label)
xgb.DMatrix.save(dtrain, 'xgb.DMatrix.data')
dtrain <- xgb.DMatrix('xgb.DMatrix.data')
if (file.exists('xgb.DMatrix.data')) file.remove('xgb.DMatrix.data')

[Package xgboost version 1.0.0.1 Index]