xgb.dump {xgboost}	R Documentation

Dump an xgboost model in text format.

Description

Dump an xgboost model in text format.

Usage

xgb.dump(model, fname = NULL, fmap = "", with_stats = FALSE,
  dump_format = c("text", "json"), ...)

Arguments

model
the model object.

fname
the name of the text file where to save the model text dump. If not provided or set to NULL, the model is returned as a character vector.

fmap
feature map file representing feature types. Detailed description could be found at https://github.com/dmlc/xgboost/wiki/Binary-Classification#dump-model. See demo/ for walkthrough example in R, and https://github.com/dmlc/xgboost/blob/master/demo/data/featmap.txt for example Format.

with_stats
whether to dump some additional statistics about the splits. When this option is on, the model dump contains two additional values: gain is the approximate loss function gain we get in each split; cover is the sum of second order gradient in each node.

dump_format
either 'text' or 'json' format could be specified.

...
currently not used

Value

If fname is not provided or set to NULL the function will return the model as a character vector. Otherwise it will return TRUE.

Examples

data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
train <- agaricus.train
test <- agaricus.test
bst <- xgboost(data = train$data, label = train$label, max_depth = 2, 
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")
# save the model in file 'xgb.model.dump'
dump_path = file.path(tempdir(), 'model.dump')
xgb.dump(bst, dump_path, with_stats = TRUE)

# print the model without saving it to a file
print(xgb.dump(bst, with_stats = TRUE))

# print in JSON format:
cat(xgb.dump(bst, with_stats = TRUE, dump_format='json'))

[Package xgboost version 1.0.0.1 Index]