xgb.gblinear.history {xgboost}	R Documentation

Extract gblinear coefficients history.

Description

A helper function to extract the matrix of linear coefficients' history from a gblinear model created while using the cb.gblinear.history() callback.

Usage

xgb.gblinear.history(model, class_index = NULL)

Arguments

model
either an xgb.Booster or a result of xgb.cv(), trained using the cb.gblinear.history() callback.

class_index
zero-based class index to extract the coefficients for only that specific class in a multinomial multiclass model. When it is NULL, all the coefficients are returned. Has no effect in non-multiclass models.

Value

For an xgb.train result, a matrix (either dense or sparse) with the columns corresponding to iteration's coefficients (in the order as xgb.dump() would return) and the rows corresponding to boosting iterations.

For an xgb.cv result, a list of such matrices is returned with the elements corresponding to CV folds.

[Package xgboost version 1.0.0.1 Index]