xgb.load {xgboost}	R Documentation

Load xgboost model from binary file

Description

Load xgboost model from the binary model file.

Usage

xgb.load(modelfile)

Arguments

modelfile
the name of the binary input file.

Details

The input file is expected to contain a model saved in an xgboost-internal binary format using either xgb.save or cb.save.model in R, or using some appropriate methods from other xgboost interfaces. E.g., a model trained in Python and saved from there in xgboost format, could be loaded from R.

Note: a model saved as an R-object, has to be loaded using corresponding R-methods, not xgb.load.

Value

An object of xgb.Booster class.

See Also

xgb.save, xgb.Booster.complete.

Examples

data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
train <- agaricus.train
test <- agaricus.test
bst <- xgboost(data = train$data, label = train$label, max_depth = 2, 
               eta = 1, nthread = 2, nrounds = 2,objective = "binary:logistic")
xgb.save(bst, 'xgb.model')
bst <- xgb.load('xgb.model')
if (file.exists('xgb.model')) file.remove('xgb.model')
pred <- predict(bst, test$data)

[Package xgboost version 1.0.0.1 Index]