xgb.parameters<- {xgboost}	R Documentation

Accessors for model parameters.

Description

Only the setter for xgboost parameters is currently implemented.

Usage

xgb.parameters(object) <- value

Arguments

object
Object of class xgb.Booster or xgb.Booster.handle.

value
a list (or an object coercible to a list) with the names of parameters to set and the elements corresponding to parameter values.

Details

Note that the setter would usually work more efficiently for xgb.Booster.handle than for xgb.Booster, since only just a handle would need to be copied.

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train

bst <- xgboost(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")

xgb.parameters(bst) <- list(eta = 0.1)

[Package xgboost version 1.0.0.1 Index]