xgb.serialize {xgboost}	R Documentation

Serialize the booster instance into R's raw vector. The serialization method differs from xgb.save.raw as the latter one saves only the model but not parameters. This serialization format is not stable across different xgboost versions.

Description

Serialize the booster instance into R's raw vector. The serialization method differs from xgb.save.raw as the latter one saves only the model but not parameters. This serialization format is not stable across different xgboost versions.

Usage

xgb.serialize(booster)

Arguments

booster
the booster instance

Examples

data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
train <- agaricus.train
test <- agaricus.test
bst <- xgboost(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2,objective = "binary:logistic")
raw <- xgb.serialize(bst)
bst <- xgb.unserialize(raw)

[Package xgboost version 1.7.3.1 Index]