xgb.unserialize {xgboost}	R Documentation

Load the instance back from xgb.serialize

Description

Load the instance back from xgb.serialize

Usage

xgb.unserialize(buffer, handle = NULL)

Arguments

buffer
the buffer containing booster instance saved by xgb.serialize

handle
An xgb.Booster.handle object which will be overwritten with the new deserialized object. Must be a null handle (e.g. when loading the model through 'readRDS'). If not provided, a new handle will be created.

Value

An xgb.Booster.handle object.

[Package xgboost version 1.7.3.1 Index]