xgboost-deprecated {xgboost}	R Documentation

Deprecation notices.

Description

At this time, some of the parameter names were changed in order to make the code style more uniform. The deprecated parameters would be removed in the next release.

Details

To see all the current deprecated and new parameters, check the xgboost:::depr_par_lut table.

A deprecation warning is shown when any of the deprecated parameters is used in a call. An additional warning is shown when there was a partial match to a deprecated parameter (as R is able to partially match parameter names).

[Package xgboost version 1.0.0.1 Index]