Accessors for serializable attributes of a model.

These methods allow to manipulate the key-value attribute strings of an xgboost model.

xgb.attr(object, name)

xgb.attr(object, name) <- value

xgb.attributes(object)

xgb.attributes(object) <- value

Arguments

object
Object of class xgb.Booster or xgb.Booster.handle.

name
a non-empty character string specifying which attribute is to be accessed.

value
a value of an attribute for xgb.attr<-; for xgb.attributes<- it's a list (or an object coercible to a list) with the names of attributes to set and the elements corresponding to attribute values. Non-character values are converted to character. When attribute value is not a scalar, only the first index is used. Use NULL to remove an attribute.

Value

xgb.attr returns either a string value of an attribute or NULL if an attribute wasn't stored in a model.

xgb.attributes returns a list of all attribute stored in a model or NULL if a model has no stored attributes.

Details

The primary purpose of xgboost model attributes is to store some meta-data about the model. Note that they are a separate concept from the object attributes in R. Specifically, they refer to key-value strings that can be attached to an xgboost model, stored together with the model's binary representation, and accessed later (from R or any other interface). In contrast, any R-attribute assigned to an R-object of xgb.Booster class would not be saved by xgb.save because an xgboost model is an external memory object and its serialization is handled externally. Also, setting an attribute that has the same name as one of xgboost's parameters wouldn't change the value of that parameter for a model. Use xgb.parameters<- to set or change model parameters.

The attribute setters would usually work more efficiently for xgb.Booster.handle than for xgb.Booster, since only just a handle (pointer) would need to be copied. That would only matter if attributes need to be set many times. Note, however, that when feeding a handle of an xgb.Booster object to the attribute setters, the raw model cache of an xgb.Booster object would not be automatically updated, and it would be user's responsibility to call xgb.save.raw to update it.

The xgb.attributes<- setter either updates the existing or adds one or several attributes, but it doesn't delete the other existing attributes.

Examples

data(agaricus.train, package='xgboost')
train <- agaricus.train

bst <- xgboost(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")

#> [1]	train-error:0.046522 
#> [2]	train-error:0.022263 

xgb.attr(bst, "my_attribute") <- "my attribute value"
print(xgb.attr(bst, "my_attribute"))

#> [1] "my attribute value"

xgb.attributes(bst) <- list(a = 123, b = "abc")

xgb.save(bst, 'xgb.model')

#> [1] TRUE

bst1 <- xgb.load('xgb.model')
if (file.exists('xgb.model')) file.remove('xgb.model')

#> [1] TRUE

print(xgb.attr(bst1, "my_attribute"))

#> [1] "my attribute value"

print(xgb.attributes(bst1))

#> $a
#> [1] "123"
#> 
#> $b
#> [1] "abc"
#> 
#> $my_attribute
#> [1] "my attribute value"
#> 
#> $niter
#> [1] "1"
#> 

# deletion:
xgb.attr(bst1, "my_attribute") <- NULL
print(xgb.attributes(bst1))

#> $a
#> [1] "123"
#> 
#> $b
#> [1] "abc"
#> 
#> $niter
#> [1] "1"
#> 

xgb.attributes(bst1) <- list(a = NULL, b = NULL)
print(xgb.attributes(bst1))

#> $niter
#> [1] "1"
#>