xgb.save {xgboost}	R Documentation

Save xgboost model to binary file

Description

Save xgboost model to a file in binary format.

Usage

xgb.save(model, fname)

Arguments

model
model object of xgb.Booster class.

fname
name of the file to write.

Details

This methods allows to save a model in an xgboost-internal binary format which is universal among the various xgboost interfaces. In R, the saved model file could be read-in later using either the xgb.load function or the xgb_model parameter of xgb.train.

Note: a model can also be saved as an R-object (e.g., by using readRDS or save). However, it would then only be compatible with R, and corresponding R-methods would need to be used to load it. Moreover, persisting the model with readRDS or save) will cause compatibility problems in future versions of XGBoost. Consult a-compatibility-note-for-saveRDS-save to learn how to persist models in a future-proof way, i.e. to make the model accessible in future releases of XGBoost.

See Also

xgb.load, xgb.Booster.complete.

Examples

data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
train <- agaricus.train
test <- agaricus.test
bst <- xgboost(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2,objective = "binary:logistic")
xgb.save(bst, 'xgb.model')
bst <- xgb.load('xgb.model')
if (file.exists('xgb.model')) file.remove('xgb.model')
pred <- predict(bst, test$data)

[Package xgboost version 1.7.3.1 Index]